refactor: StaffItemクラスを追加
