+++ lib/features/staff/ui/staff_section.dart
@@ -1,13 +1,10 @@
 import 'package:confwebsite2023/core/components/responsive_widget.dart';
 import 'package:confwebsite2023/core/components/section_header.dart';
-import 'package:confwebsite2023/core/theme.dart';
-import 'package:confwebsite2023/features/staff/data/staff.dart';
+import 'package:confwebsite2023/core/theme/app_text_style.dart';
 import 'package:confwebsite2023/features/staff/data/staff_provider.dart';
-import 'package:confwebsite2023/features/staff/ui/sns_icon.dart';
+import 'package:confwebsite2023/features/staff/ui/staff_item.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:transparent_image/transparent_image.dart';
-import 'package:url_launcher/link.dart';
 
 class StaffSection extends StatelessWidget {
   const StaffSection({super.key});
@@ -83,21 +80,21 @@ class _StaffList extends ConsumerWidget {
                                   .map<Widget>(
                                     (e) => SizedBox(
                                       width: itemWidth,
-                                      child: _StaffItem(staff: e),
+                                      child: StaffItem(staff: e),
                                     ),
                                   )
                                   .toList()
                                   .joinElement(
-                                    const SizedBox(width: _StaffItem.spacing),
+                                    const SizedBox(width: StaffItem.spacing),
                                   ),
                             ),
                           ),
                         )
                         .toList()
                         .joinElement(
-                          const SizedBox(height: _StaffItem.spacing),
+                          const SizedBox(height: StaffItem.spacing),
                         ),
-                    Spaces.vertical_200,
+                    const SizedBox(height: 165),
                   ],
                 ),
               );
@@ -111,104 +108,6 @@ class _StaffList extends ConsumerWidget {
       loading: () {
         return const CircularProgressIndicator.adaptive();
       },
-    );
-  }
-}
-
-class _StaffItem extends StatelessWidget {
-  const _StaffItem({
-    required this.staff,
-  });
-  final Staff staff;
-
-  static const spacing = 16.0;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final icon = SizedBox(
-      height: 120,
-      width: 120,
-      child: FittedBox(
-        child: ClipOval(
-          child: FadeInImage(
-            fit: BoxFit.cover,
-            image: NetworkImage(staff.image.src),
-            placeholder: MemoryImage(kTransparentImage),
-            imageErrorBuilder: (_, __, ___) => const FittedBox(
-              child: Icon(
-                Icons.error,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    final snsIcons = Wrap(
-      alignment: WrapAlignment.center,
-      children: staff.sns
-          .map(
-            (e) => Link(
-              target: LinkTarget.blank,
-              uri: Uri.tryParse(e.type.prefixUrl + e.value),
-              builder: (_, followLink) => SnsIcon(
-                snsType: e.type,
-                size: 32,
-                padding: EdgeInsets.zero,
-                iconColor: theme.colorScheme.onPrimaryContainer,
-                onPressed: followLink,
-              ),
-            ),
-          )
-          .toList(),
-    );
-
-    return Card(
-      elevation: 0,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-      ),
-      color: theme.colorScheme.secondaryContainer,
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(
-              top: 16,
-              left: 24,
-              right: 24,
-            ),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                icon,
-                Spaces.vertical_16,
-                Text(
-                  staff.displayName,
-                  style: theme.textTheme.titleLarge!.copyWith(
-                    color: baselineColorScheme.ref.primary.primary100,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-                Spaces.vertical_8,
-                Text(
-                  staff.introduction,
-                  style: theme.textTheme.bodyLarge!.copyWith(
-                    color: baselineColorScheme.ref.primary.primary80,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-              ],
-            ),
-          ),
-          const Spacer(),
-          Spaces.vertical_16,
-          snsIcons,
-          Spaces.vertical_16,
-        ],
-      ),
     );
   }
 }
